--- a/gulpfile.js
+++ b/gulpfile.js
@@ -182,6 +182,18 @@ function declareExtensionVersionAlias(name, version, lastestVersion, hasCss) {
   }
 }
 
+
+/**
+ * Logs a build step to the console, unless we are on Travis, where we don't.
+ * @param {string} stepName Name of an action, like 'Bundling' or 'Compiling'
+ * @param {string} targetName Name of the target, like a filename or path
+ */
+function logBuildStep(stepName, targetName) {
+  if(!process.env.TRAVIS) {
+    $$.util.log(stepName, $$.util.colors.cyan(targetName));
+  }
+}
+
 /**
  * Build all the AMP extensions
  *
@@ -351,9 +363,7 @@ function compile(watch, shouldMinify, opt_preventRemoveAndMakeDir,
  * @return {!Promise}
  */
 function compileCss() {
-  if (!process.env.TRAVIS) {
-    $$.util.log('Recompiling CSS in amp.css');
-  }
+  logBuildStep('Recompiling CSS in', 'amp.css');
   return jsifyCssAsync('css/amp.css').then(function(css) {
     return toPromise(gulp.src('css/**.css')
           .pipe($$.file('css.js', 'export const cssText = ' +
@@ -372,9 +382,7 @@ function compileCss() {
  * @return {!Promise}
  */
 function copyCss() {
-  if (!process.env.TRAVIS) {
-    $$.util.log('Copying build/css/v0.css to dist/v0.css');
-  }
+  logBuildStep('Copying', 'build/css/v0.css to dist/v0.css');
   fs.copySync('build/css/v0.css', 'dist/v0.css');
   return toPromise(gulp.src('build/css/amp-*.css')
       .pipe(gulp.dest('dist/v0')));
@@ -448,9 +456,7 @@ function buildExtension(name, version, hasCss, options, opt_extraGlobs) {
   if (hasCss) {
     mkdirSync('build');
     mkdirSync('build/css');
-    if (!process.env.TRAVIS) {
-      $$.util.log('Recompiling CSS in ' + name);
-    }
+    logBuildStep('Recompiling CSS in', name);
     return jsifyCssAsync(path + '/' + name + '.css').then(function(css) {
       var jsCss = 'export const CSS = ' + JSON.stringify(css) + ';\n';
       var jsName = 'build/' + name + '-' + version + '.css.js';
@@ -479,9 +485,7 @@ function buildExtension(name, version, hasCss, options, opt_extraGlobs) {
  * @return {!Promise}
  */
 function buildExtensionJs(path, name, version, options) {
-  if (!process.env.TRAVIS) {
-    $$.util.log('Bundling ' + name);
-  }
+  logBuildStep('Bundling', name);
   var filename = options.filename || name + '.js';
   if (options.loadPriority && options.loadPriority != 'high') {
     throw new Error('Unsupported loadPriority: ' + options.loadPriority);
@@ -645,10 +649,7 @@ function checkTypes() {
  * @return {!Promise}
  */
 function thirdPartyBootstrap(input, outputName, shouldMinify) {
-  if (!process.env.TRAVIS) {
-    $$.util.log('Processing ' + input);
-  }
-
+  logBuildStep('Processing', input);
   if (!shouldMinify) {
     return toPromise(gulp.src(input)
         .pipe(gulp.dest('dist.3p/current')));
@@ -722,10 +723,7 @@ function appendToCompiledFile(srcFilename, destFilePath) {
 function compileJs(srcDir, srcFilename, destDir, options) {
   options = options || {};
   if (options.minify) {
-    if (!process.env.TRAVIS) {
-      $$.util.log('Minifying ' + srcFilename);
-    }
-
+    logBuildStep('Minifying', srcFilename);
     return closureCompile(
         srcDir + srcFilename, destDir, options.minifiedName, options)
         .then(function() {
@@ -774,17 +772,13 @@ function compileJs(srcDir, srcFilename, destDir, options) {
       .pipe(lazywrite())
       .on('end', function() {
         appendToCompiledFile(srcFilename, destDir + '/' + destFilename);
-        if (!process.env.TRAVIS) {
-          $$.util.log('Compiled ' + srcFilename);
-        }
+        logBuildStep('Compiled', srcFilename);
       }));
   }
 
   if (options.watch) {
     bundler.on('update', function() {
-      if (!process.env.TRAVIS) {
-        $$.util.log('-> bundling ' + srcDir + '...');
-      }
+      logBuildStep('-> bundling', srcDir + '...');
       rebundle();
       // Touch file in unit test set. This triggers rebundling of tests because
       // karma only considers changes to tests files themselves re-bundle
@@ -802,11 +796,8 @@ function compileJs(srcDir, srcFilename, destDir, options) {
  * @param {!Object} options
  */
 function buildExperiments(options) {
+  logBuildStep('Bundling', 'experiments.html/js');
   options = options || {};
-  if (!process.env.TRAVIS) {
-    $$.util.log('Bundling experiments.html/js');
-  }
-
   var path = 'tools/experiments';
   var htmlPath = path + '/experiments.html';
   var jsPath = path + '/experiments.js';
@@ -828,9 +819,7 @@ function buildExperiments(options) {
   }
 
   // Build HTML.
-  if (!process.env.TRAVIS) {
-    $$.util.log('Processing ' + htmlPath);
-  }
+  logBuildStep('Processing', htmlPath);
   var html = fs.readFileSync(htmlPath, 'utf8');
   var minHtml = html.replace('/dist.tools/experiments/experiments.js',
       `https://${hostname}/v0/experiments.js`);
@@ -873,11 +862,8 @@ function buildLoginDone(options) {
  * @param {!Object} options
  */
 function buildLoginDoneVersion(version, options) {
+  logBuildStep('Bundling', 'amp-login-done.html/js');
   options = options || {};
-  if (!process.env.TRAVIS) {
-    $$.util.log('Bundling amp-login-done.html/js');
-  }
-
   var path = 'extensions/amp-access/' + version + '/';
   var htmlPath = path + 'amp-login-done.html';
   var jsPath = path + 'amp-login-done.js';
@@ -899,9 +885,7 @@ function buildLoginDoneVersion(version, options) {
   }
 
   // Build HTML.
-  if (!process.env.TRAVIS) {
-    $$.util.log('Processing ' + htmlPath);
-  }
+  logBuildStep('Processing', htmlPath);
   var html = fs.readFileSync(htmlPath, 'utf8');
   var minHtml = html.replace(
       '../../../dist/v0/amp-login-done-' + version + '.max.js',
@@ -939,11 +923,8 @@ function buildLoginDoneVersion(version, options) {
  * @param {!Object} options
  */
 function buildAlp(options) {
+  logBuildStep('Bundling', 'alp.js');
   options = options || {};
-  if (!process.env.TRAVIS) {
-    $$.util.log('Bundling alp.js');
-  }
-
   return compileJs('./ads/alp/', 'install-alp.js', './dist/', {
     toName: 'alp.max.js',
     watch: options.watch,
@@ -960,11 +941,8 @@ function buildAlp(options) {
  * @param {!Object} options
  */
 function buildExaminer(options) {
+  logBuildStep('Bundling', 'examiner.js');
   options = options || {};
-  if (!process.env.TRAVIS) {
-    $$.util.log('Bundling examiner.js');
-  }
-
   return compileJs('./src/examiner/', 'examiner.js', './dist/', {
     toName: 'examiner.max.js',
     watch: options.watch,
@@ -981,11 +959,8 @@ function buildExaminer(options) {
  * @param {!Object} options
  */
 function buildSw(options) {
-  if (!process.env.TRAVIS) {
-    $$.util.log('Bundling sw.js');
-  }
+  logBuildStep('Bundling', 'sw.js');
   var opts = Object.assign({}, options);
-
   return Promise.all([
     // The service-worker script loaded by the browser.
     compileJs('./src/service-worker/', 'shell.js', './dist/', {
@@ -1015,11 +990,8 @@ function buildSw(options) {
  * @param {!Object} options
  */
 function buildWebWorker(options) {
-  if (!process.env.TRAVIS) {
-    $$.util.log('Bundling ww.js');
-  }
+  logBuildStep('Bundling', 'ww.js');
   var opts = Object.assign({}, options);
-
   return compileJs('./src/web-worker/', 'web-worker.js', './dist/', {
     toName: 'ww.max.js',
     minifiedName: 'ww.js',