Prior to this PR, `gulp build --css-only` would take ~ 1 minute to run, and end up building a bunch of js files in addition to compiling css.

After this PR, `gulp build --css-only` takes ~ 0.5 seconds to run (a 100x speed up), and the only output files in `build/` after running it are .css, and .css.js files.

Resulting speeds on my workstation:
`gulp build --css-only`: ~500ms
`gulp build`: ~1m 45s
`gulp dist --fortesting`: ~3m 30s

Fixes #9640