@rsimha-amp 2017-06-06
We don't log steps on Travis because the logs will become more verbose than we want. Currently, build is only a small part of the Travis PR logs.