@rsimha-amp 2017-06-06
@choumx, I've removed the spurious compileCss from compile. Feel free to take another look.